pyproject.toml
[project]
name = "medicalassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.1",
    "fastapi>=0.132.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.22.0",
    "pandas>=3.0.1",
    "pinecone-client>=6.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.7.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]
